// 2019/website/src/widgets/common/interfaces.d.ts
export type AriaPropertyObject = {
	[key: string]: string;
};

export interface CustomAriaProperties {
	aria?: AriaPropertyObject;
}

export interface Focus extends FocusEvent {
	readonly value?: string;
}
export interface Mouse extends MouseEvent {
	readonly previousScreenX: number;
	readonly previousScreenY: number;
	readonly movementX: number;
	readonly movementY: number;
	readonly relatedTarget: any;
	readonly targetKey: string | null;
}
export interface Input extends UIEvent {
	readonly value: string;
	readonly data: string;
	readonly dataTransfer: DataTransfer;
	readonly inputType: 'insertText' | 'insertCompositionText' | 'insertFromComposition' |
	'insertFromPaste' | 'insertFromDrop' | 'insertReplacementText' | 'insertFromYank'|
	'formatSetBlockTextDirection' | 'formatSetInlineTextDirection' | 'formatBackColor'|
	'formatFontColor' | 'formatFontName' | 'insertLink' | 'historyUndo' | 'historyRedo';
  readonly isComposing: boolean;
}
export interface Toggle extends MouseEvent {
	readonly value: string;
	readonly checked: boolean;
}
export interface Keyboard extends KeyboardEvent {
	readonly which: number;
	readonly key: Keys;
	readonly isStart: boolean;
	readonly isEnd: boolean;
}
// ...
export interface LabeledProperties {
	labelAfter?: boolean;
	labelHidden?: boolean;
	labelStatic?: boolean;
	label?: DNode;
	helperText?: string;
}
// ...
export interface GenericInputProperties {
	disabled?: boolean;
	widgetId?: string;
	valid?: { valid?: boolean; message?: string; } | boolean;

	name?: string;
	readOnly?: boolean;
	required?: boolean;
}

export interface WidgetEventProperties {
	onHover?: (evt: MouseEvent) => void;
	onClick?(evt: MouseEvent): void;
}
// ...
export interface CheckEventProperties extends WidgetEventProperties {
	onFocus?(evt: FocusEvent): void;
	onBlur?(evt: FocusEvent): void;
	onChange?(evt: Toggle): void;
}
// ...
export interface KeyEventProperties {
	onKeyDown?(event: Keyboard, preventDefault: () => void): void;
	onKeyPress?(event: Keyboard, preventDefault: () => void): void;
	onKeyUp?(event: Keyboard, preventDefault: () => void): void;
}

/**
 * @type PointerEventProperties
 * @property onMouseDown    Called on the input's mousedown event
 * @property onMouseUp      Called on the input's mouseup event
 * @property onTouchStart   Called on the input's touchstart event
 * @property onTouchEnd     Called on the input's touchend event
 * @property onTouchCancel  Called on the input's touchcancel event
 */
export interface PointerEventProperties {
	onMouseDown?(event: MouseEvent): void;
	onMouseUp?(event: MouseEvent): void;
	onTouchStart?(event: TouchEvent): void;
	onTouchEnd?(event: TouchEvent): void;
	onTouchCancel?(event: TouchEvent): void;
}
// ...
export interface RedaktorProperties extends ThemedProperties {
  required?: boolean;
  disabled?: boolean;
  readOnly?: boolean;
  invalid?: boolean;
  responsive?: boolean;
	size?: Sizes | undefined;
	schema?: any; // MaterialSchema | keyof typeof MaterialSchema; // TODO
  filled?: boolean;
  outlined?: boolean;
  shaped?: boolean;
}
export interface RedaktorBaseCSS {
  /* only ENUMS Material and Size TODO */
  [index:string]: string;
  // FIXME TS next https://github.com/Microsoft/TypeScript/pull/23592
}
export interface RedaktorSizesCSS extends RedaktorBaseCSS {
  smallUI: string;
  defaultUI: string;
  mediumUI: string;
  largeUI: string;
  smallTypo: string;
  defaultTypo: string;
  mediumTypo: string;
  largeTypo: string;
  responsive: string;
}
export interface RedaktorSchemaCSS extends RedaktorBaseCSS {
  parentSchema: string;
  hasSchema?: any;
}
export interface RedaktorDisabledCSS extends RedaktorBaseCSS {
  enabled: string;
  disabled: string;
  readonly?: any;
}
export interface RedaktorValidCSS extends RedaktorBaseCSS {
  valid: string;
  invalid: string;
}
export interface RedaktorStyleCSS extends RedaktorBaseCSS {
  filled: string;
  outlined: string;
  shaped: string;
}
export interface RedaktorPositionCSS extends RedaktorBaseCSS {
	left: string;
	center: string;
	right: string;
	top: string;
	middle: string;
	bottom: string;
}

export type RedaktorCSS = RedaktorDisabledCSS & RedaktorValidCSS & RedaktorSchemaCSS;